Update Bomberman
PixelPad Classes dont take in parameters, unless they are declared in the Functions section, in which they need a regular PixelPad class just for the sprite. The BG doesn't really need parameters anyway, and the sprite can be defined like so after the initialization of the class. Alternatively, we could put: "self.sprite = sprite('Background.jpg')" in the BG class' start function, and get rid of the definition of the BG's sprite in the game start function altogether.